requirements.txt
gunicorn==19.6.0
honcho==0.7.1
aiohttp==1.0.1
cchardet==1.0.0
aiodns==1.1.1
PyYAML==3.12
jsonschema==2.5.1